% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_prov_state}
\alias{assign_prov_state}
\title{Categorize polygon by Province/State}
\usage{
assign_prov_state(
  strata_map,
  min_overlap = 0.75,
  plot = FALSE,
  keep_spatial = TRUE
)
}
\arguments{
\item{strata_map}{sf data frame. Strata polygons to be categorized.}

\item{min_overlap}{Numeric. The minimum proportion of overlap between a
stratum polygon and a Province or State. Overlap below this proportion will
raise warnings.}

\item{plot}{Logical. Whether to plot how polygons were assigned to Provinces
or States}

\item{keep_spatial}{Logical. Whether the output should be a spatial data
frame or not.}
}
\value{
(Spatial) data frame with strata assigned to Province/State.
}
\description{
Categorizes custom stratification polygons by province or state if possible.
This can be useful for calculating regional indices (\code{generate_indices()}) or
trends (\code{generate_trends()}) on a custom stratification, or if you want to
create a geofaceted plot (\code{plot_geofacet()}).
}
\examples{
# Demonstration of why we can't divide BCR by Provinces and States!
map <- load_map("bcr")
assign_prov_state(map, plot = TRUE)

# Use custom stratification, using sf map object
# e.g. with WBPHS stratum boundaries 2019
# available: https://ecos.fws.gov/ServCat/Reference/Profile/142628

\dontrun{
map <- sf::read_sf("../WBPHS_Stratum_Boundaries_2019") \%>\%
  rename(strata_name = STRAT) # expects this column
s <- assign_prov_state(map, plot = TRUE)
# Some don't divide nicely, we could try a different min_overlap

s <- assign_prov_state(map, min_overlap = 0.6, plot = TRUE)
}


}
\seealso{
Other helper functions: 
\code{\link{load_map}()},
\code{\link{search_species}()}
}
\concept{helper functions}
