% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_strata}
\alias{bbs_strata}
\title{List of included strata}
\format{
\subsection{\code{bbs_strata}}{

A list of 6 data frames

Contains
\code{bbs} - New as of version 1.1.3.1
Intersection of States-Provinces-Territories with the updated (as of 2025)
Bird Conservation Regions (BCRs). Updated BCRs include changes
made to the Northern BCRs. BCR "numbers" are now alpha-numeric to account
for new regions that are subdivisions of previous numbered BCRs. For example,
BCRs 6S and 6N are southern and northern subregions of the previous BCR 6.
\code{bbs_usgs} - Intersection of States-Provinces-Territories with the previous
version of the Bird Conservation Regions (BCRs)
\code{bbs_cws}- Same as bbs_usgs, with two modifications to suit Canadian context:
Intersection of States-Provinces-Territories with the previous version
of the Bird Conservation Regions (BCRs), and BCR 7 is treated as a single
stratum (pooled across all provinces and territories), and Prince Edward
Island is combined with Nova Scotia - pooled across the two provinces
\code{bcr} - Updated (2025) version of the Bird Conservation Regions (BCRs).
\code{latlong} - 1-degree latitude * longitude grid. The design stratification of
the BBS.
\code{prov_state} - Canadian provinces/territories and states of the continental
United States of America - excludes Hawaii (there are no BBS routes in HI)
}
}
\usage{
bbs_strata
}
\description{
List of strata included in bbsBayes2. Each list item contains a data frame
describing the strata for that stratification (name, area, country, etc.)
}
\examples{
bbs_strata[["bbs_cws"]]
bbs_strata[["latlon"]]
}
\keyword{datasets}
