% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsBayes2-defunct.R
\name{bbsBayes2-defunct}
\alias{bbsBayes2-defunct}
\title{bbsBayes2 defunct functions}
\arguments{
\item{...}{Original function arguments}
}
\description{
\subsection{Superseded}{

No superseded functions
}

\subsection{No longer relevant}{

No non-relevant functions#'
}
}
\seealso{
\link{bbsBayes2-deprecated}
}
