% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsBayes2-deprecated.R
\name{bbsBayes2-deprecated}
\alias{bbsBayes2-deprecated}
\title{bbsBayes2 deprecated functions}
\arguments{
\item{...}{Original function arguments}
}
\description{
No deprecated functions
}
\seealso{
\link{bbsBayes2-defunct}
}
