% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsBayes2-package.R
\docType{package}
\name{bbsBayes2-package}
\alias{bbsBayes2}
\alias{bbsBayes2-package}
\title{bbsBayes2: Hierarchical Bayesian Analysis of North American BBS Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The North American Breeding Bird Survey (BBS) is a long-running program that seeks to monitor the status and trends of the breeding birds in North America. Since its start in 1966, the BBS has accumulated over 50 years of data for over 500 species of North American Birds. Given the temporal and spatial structure of the data, hierarchical Bayesian models are used to assess the status and trends of these 500+ species of birds. 'bbsBayes2' allows you to perform hierarchical Bayesian analysis of BBS data. You can run a full model analysis for one or more species that you choose, or you can take more control and specify how the data should be stratified, prepared for 'Stan', or modelled.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbsBayes/bbsBayes2}
  \item \url{https://bbsbayes.github.io/bbsBayes2/}
  \item Report bugs at \url{https://github.com/bbsBayes/bbsBayes2/issues}
}

}
\author{
\strong{Maintainer}: Brandon P.M. Edwards \email{brandonedwards3@cmail.carleton.ca}

Authors:
\itemize{
  \item Adam C. Smith \email{adam.smith@ec.gc.ca}
  \item Steffi LaZerte \email{steffi@steffi.ca}
}

}
\keyword{internal}
