% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equal_area_crs}
\alias{equal_area_crs}
\title{equal_area_crs}
\format{
\subsection{\code{equal_area_crs}}{

An object of class \code{crs}
}
}
\usage{
equal_area_crs
}
\description{
Introduced in version 1.1.2.
The specific coordinate reference system (crs) used for the included strata maps.
North_America_Albers_Equal_Area_Conic, projection that provides good visualisation
of the main areas covered by the North American Breeding Bird Survey
}
\examples{
equal_area_crs
}
\keyword{datasets}
