% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{fetch_bbs_data}
\alias{fetch_bbs_data}
\title{Fetch Breeding Bird Survey dataset}
\usage{
fetch_bbs_data(
  level = "state",
  release = 2025,
  force = FALSE,
  quiet = FALSE,
  compression = "none",
  include_unacceptable = FALSE
)
}
\arguments{
\item{level}{Character. Which type of BBS counts to use, "state" or "stop".
Default "state".}

\item{release}{Numeric. Which yearly release to use, 2022 (including data
through 2021 field season) or 2020 (including data through 2019). Default
2022.}

\item{force}{Logical. Should pre-exising BBS data be overwritten? Default
FALSE.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}

\item{compression}{Character. What compression should be used to save data?
Default is \code{none} which takes up the most space but is the fastest to
load. Must be one of \code{none}, \code{gz}, \code{bz2}, or \code{xz} (passed to
\code{readr::write_rds()}'s \code{compress} argument).}

\item{include_unacceptable}{Logical. DO NOT change this unless you are very
confident that you want these data. Should the package retain the BBS data
collected on non-standard BBS routes (including some with fewer than 50
stops), or outside of the acceptable weather, season, and times-of-day
structure of the BBS survey design. Default FALSE for good reason.
If you change this to TRUE and thereby decide you want to include
the BBS data that do not fit the BBS structured survey design, be sure you
understand the information in at least the following columns of the route
survey info: RouteTypeID, RouteTypeDetailID, RPID, RunType, and the rest
of the information in the data release meta data.}
}
\description{
Fetch and download Breeding Bird Survey data from the United States
Geological Survey (USGS) FTP site. This is the raw data that is uploaded to
the site before any analyses are performed. Users can download different
types (\code{state}, \code{stop}) and different releases (currently \code{2020}, \code{2022},
\code{2023}, \code{2024}, and \code{2025}).
}
\details{
Users will be asked before saving the BBS data to a package-specific
directory created on their computer. Before downloading any data, users must
thoroughly read through the USGS terms and conditions for that data and enter
the word "yes" to agree.

BBS \code{state} level counts provide counts beginning in 1966, aggregated in five
bins, each of which contains cumulative counts from 10 of the 50 stops along
a route. In contrast, BBS \code{stop} level counts provides stop-level data
beginning in 1997, which includes counts for each individual stop along
routes. \strong{Note that stop-level data is not currently supported by the
modelling utilities in bbsBayes2.}

There are multiple releases for each type of data, \code{2020}, \code{2022}, \code{2023},
\code{2024} and \code{2025}.
By default all functions use the most recent release unless otherwise
specified. For example, the \code{release} argument in \code{stratify()} can be
changed to \code{2020} to use the 2020 release of state-level counts.

bbsBayes2 by default removes observations from routes that were not
established following the stratified random design, the handful of routes
that are on water (not on roadsides), and any survey that was not conducted
within acceptable survey conditions (high winds, heavy precipitation,
outside of the acceptable time windows).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

fetch_bbs_data(force = TRUE)
fetch_bbs_data(level = "stop", force = TRUE)
fetch_bbs_data(release = 2020, force = TRUE)
fetch_bbs_data(release = 2020, level = "stop", force = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other BBS data functions: 
\code{\link{have_bbs_data}()},
\code{\link{load_bbs_data}()},
\code{\link{remove_cache}()}
}
\concept{BBS data functions}
