% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_summary}
\alias{get_summary}
\title{Return the \code{cmdstanr} summary}
\usage{
get_summary(model_output, variables = NULL)
}
\arguments{
\item{model_output}{List. Model output generated by \code{run_model()}.}

\item{variables}{Character vector. Specific variables (e.g.,
\code{"strata_raw[1]"}) or variable types (e.g., \code{"strata_raw"}) for which to
calculate metrics. If \code{NULL} (default) all variables are returned.}
}
\value{
A data frame of model summary statistics, matching the output of
\code{cmdstanr::summary()} plus one additional column \code{variable_type} that
identifies categories of parameters (e.g., \code{n} for all of the annual
indices for every stratum and year).
}
\description{
Extract and return the model summary using \code{cmdstanr::summary()}.
}
\examples{
# Temporarily suppress convergence warning for legibility
# "The ESS has been capped to avoid unstable estimates."
opts <- options(warn = -1)

# Using the example model for Pacific Wrens

get_summary(pacific_wren_model)
get_summary(pacific_wren_model, variables = "strata_raw")
get_summary(pacific_wren_model, variables = "strata_raw[9]")

# Restore warnings
options(opts)
}
\seealso{
Other model assessment functions: 
\code{\link{get_convergence}()},
\code{\link{get_model_vars}()}
}
\concept{model assessment functions}
