% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{have_cmdstan}
\alias{have_cmdstan}
\title{Check if cmdstan is installed}
\usage{
have_cmdstan()
}
\description{
Wrapper around \code{cmdstanr::cmdstan_version(error_on_NA = FALSE)} for quick
check.
}
\details{
Used internally for skipping examples and tests if no cmdstan installed.
}
