% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_map}
\alias{load_map}
\title{Load a geographic strata map}
\usage{
load_map(stratify_by = NULL, type = "strata")
}
\arguments{
\item{stratify_by}{Character. Stratification type. One of
"bbs", prov_state", "bcr", "latlong", "bbs_cws", "bbs_usgs".}

\item{type}{Character. "strata" or political map ("North America", "Canada"
or "US"/"USA"/"United States of America").}
}
\value{
sf polygon object
}
\description{
Load one of the included spatial data files (maps) as a simple features
object (\code{sf} package)
}
\examples{

# First, stratify the sample data
strat_map <- load_map(stratify_by = "bbs")

# simple plot of the map
plot(strat_map)

# or use ggplot2
library(ggplot2)
ggplot(data = strat_map) +
  geom_sf(aes(fill = strata_name), show.legend = FALSE)

}
\seealso{
Other helper functions: 
\code{\link{assign_prov_state}()},
\code{\link{search_species}()}
}
\concept{helper functions}
