% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacific_wren_model}
\alias{pacific_wren_model}
\title{Example model output}
\format{
\subsection{\code{pacific_wren_model}}{

A list output from \code{run_model()} with 4 items
\itemize{
\item \code{model_fit} - cmdstanr model output
\item \code{model_data} - list of data formatted for use in Stan modelling
\item \code{meta_data} - meta data defining the analysis
\item \code{meta_strata} - data frame listing strata meta data
\item \code{raw_data} - data frame of summarized counts
}
}
}
\usage{
pacific_wren_model
}
\description{
Example model output from running a hierarchical first difference model
on the included sample data for Pacific Wrens.
}
\examples{
# Code to replicate:

pacific_wren_model <- stratify(by = "bbs_cws", sample_data = TRUE) \%>\%
  prepare_data() \%>\%
  prepare_model(model = "first_diff", set_seed = 111) \%>\%
  run_model(chains = 2, iter_sampling = 20, iter_warmup = 20, set_seed = 111)

}
\keyword{datasets}
