% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-geofacet.R
\name{plot_geofacet}
\alias{plot_geofacet}
\title{Create geofacet plot of population trajectories by province/state}
\usage{
plot_geofacet(
  indices,
  ci_width = 0.95,
  multiple = FALSE,
  trends = NULL,
  slope = FALSE,
  add_observed_means = FALSE,
  col_viridis = FALSE
)
}
\arguments{
\item{indices}{List. Indices generated by \code{generate_indices()}.}

\item{ci_width}{Numeric. Quantile defining the width of the plotted credible
interval. Defaults to 0.95 (lower = 0.025 and upper = 0.975). Note these
quantiles need to have been precalculated in \code{generate_indices()}.}

\item{multiple}{Logical. Whether to plot multiple strata-level trajectories
within each prov/state facet. Default \code{FALSE}.}

\item{trends}{List. (Optional) Output generated by \code{generate_trends()}. If
included trajectories are coloured based on the same colour scale used in
\code{plot_map}.}

\item{slope}{Logical. If \code{trends} included, whether colours in the plot
should be based on slope trends. Default \code{FALSE}.}

\item{add_observed_means}{Logical. Whether to include points indicating the
observed mean counts. Default \code{FALSE}. Note: scale of observed means and
annual indices may not match due to imbalanced sampling among routes.}

\item{col_viridis}{Logical. Should the colour-blind-friendly "viridis"
palette be used. Default \code{FALSE}.}
}
\value{
ggplot object
}
\description{
Generate a faceted plot of population trajectories by province/state. Only
possible if indices created by \code{generate_indices()} include the \code{prov_state}
region. All geofacet plots have one facet per state/province, so if there are
multiple strata per facet, these can be plotted as separate trajectories
within each facet (\code{multiple = TRUE}).
}
\examples{
# Using the example model for Pacific Wrens...

# Generate indices
i <- generate_indices(pacific_wren_model,
                      regions = c("stratum", "prov_state"))
# Generate trends
t <- generate_trends(i)

# Now make the geofacet plot.
plot_geofacet(i, trends = t, multiple = TRUE)
plot_geofacet(i, trends = t, multiple = TRUE, col_viridis = TRUE)
plot_geofacet(i, multiple = TRUE)
plot_geofacet(i, trends = t, multiple = FALSE)
plot_geofacet(i, multiple = FALSE)

# With different ci_width, specify desired quantiles in indices
i <- generate_indices(pacific_wren_model,
                      regions = c("stratum", "prov_state"),
                      quantiles = c(0.005, 0.995))

plot_geofacet(i, multiple = FALSE, ci_width = 0.99)

}
\seealso{
Other indices and trends functions: 
\code{\link{generate_indices}()},
\code{\link{generate_trends}()},
\code{\link{plot_indices}()},
\code{\link{plot_map}()}
}
\concept{indices and trends functions}
