% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_map}
\alias{plot_map}
\title{Generate a map of trends by strata}
\usage{
plot_map(
  trends,
  slope = FALSE,
  title = TRUE,
  col_viridis = FALSE,
  strata_custom = NULL
)
}
\arguments{
\item{trends}{List. Trends generated by \code{generate_trends()}.}

\item{slope}{Logical. Whether or not to map values of the alternative trend
metric (slope of a log-linear regression) if \code{slope = TRUE} was used in
\code{generate_trends()},  through the annual indices. Default \code{FALSE}.}

\item{title}{Logical. Whether or not to include a title with species. Default
\code{TRUE}.}

\item{col_viridis}{Logical. Should the colour-blind-friendly "viridis"
palette be used. Default \code{FALSE}.}

\item{strata_custom}{(\code{sf}) Data Frame. Data frame
of modified existing stratification, or a \code{sf} spatial data frame with
polygons defining the custom stratifications. See details on strata_custom
in \code{stratify()}.}
}
\value{
a ggplot2 plot
}
\description{
\code{plot_map()} allows you to generate a colour-coded map of the percent
change in species trends for each strata.
}
\examples{
# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices
i <- generate_indices(pacific_wren_model)

# Now generate trends
t <- generate_trends(i, slope = TRUE)

# Generate the map (without slope trends)
plot_map(t)

# Generate the map (with slope trends)
plot_map(t, slope = TRUE)

# Viridis
plot_map(t, col_viridis = TRUE)

}
\seealso{
Other indices and trends functions: 
\code{\link{generate_indices}()},
\code{\link{generate_trends}()},
\code{\link{plot_geofacet}()},
\code{\link{plot_indices}()}
}
\concept{indices and trends functions}
