% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Filter for data quality}
\usage{
prepare_data(
  strata_data,
  min_year = NULL,
  max_year = NULL,
  min_n_routes = 3,
  min_max_route_years = 3,
  min_mean_route_years = 1,
  quiet = FALSE,
  assume_observer_variation_log_normal = FALSE
)
}
\arguments{
\item{strata_data}{List. Stratified data generated by \code{stratify()}}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{min_n_routes}{Numeric. Required minimum routes per strata where species
has been observed. Default 3.}

\item{min_max_route_years}{Numeric. Required minimum number of years with
non-zero observations of species on at least 1 route. Default 3. Only
retain strata with at least one route where the species was observed at
least once in this many years.}

\item{min_mean_route_years}{Numeric. Required minimum average of years per
route with the species observed. Default 1. Only retain strata where the
average number of years the species was observed per route is greater than
this value.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}

\item{assume_observer_variation_log_normal}{Logical. Default FALSE.
If FALSE, the model will generate indices that adjust only for temporal
variation in observers and so generate annual indices that are the
expected counts in a given year averaged across all observers and all
routes in the stratum. FALSE option may be useful in situations where one
or more of the 3 assumptions above are questionable, including where the
observer variation is not log-normal (e.g., heavy-tailed, or skewed). The
difference can be important because it can change the relative scaling of
annual indices among strata, which in turn influences the weight of each
strata-level trend on trends for composite regions.
If TRUE, the annual indices adjust for both spatial and temporal
variation in observers by using the same \code{retrans_obs} retransformation
factor for all strata, and so generate annual indices that are the expected
counts in a given year averaged across all routes in the stratum and assume
all observer variation is 1) noise, 2) approximately log-normally
distributed (retrans_obs = 0.5*sdobs^2), and observers are exchangeable.
This alternate TRUE is a more appropriate estimate of the uncertainty in
an expected count, generating a prediction error across all observers in the
analysis, however even small departures from the assumption of log-normally
distributed errors can introduce bias into the annual indices, relative
to the observed counts in the dataset.}
}
\value{
List of prepared (meta) data to be used for modelling and further
steps.
\itemize{
\item \code{model_data} - list of data formatted for use in Stan modelling
\item \code{meta_data} - meta data defining the analysis
\item \code{meta_strata} - data frame listing strata meta data
\item \code{raw_data} - data frame of summarized counts used to create \code{model_data}
(just formatted more nicely)
}
}
\description{
Check and filter the stratified data by minimum required samples for
modelling, and prepare data format for use by models.
}
\examples{
# Toy example with Pacific Wren sample data

# First, stratify the sample data

s <- stratify(by = "bbs_cws", sample_data = TRUE,
              use_map = FALSE)

# Prepare the stratified data for use in a model. In this
#   toy example, we will set the minimum year as 2009 and
#   maximum year as 2018, effectively only setting up to
#   model 10 years of data.

p <- prepare_data(s, min_year = 2009, max_year = 2018)
}
\seealso{
Other Data prep functions: 
\code{\link{prepare_model}()},
\code{\link{prepare_spatial}()},
\code{\link{stratify}()}
}
\concept{Data prep functions}
