% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{save_model_run}
\alias{save_model_run}
\title{Save output of run_model()}
\usage{
save_model_run(model_output, retain_csv, path = NULL, quiet = FALSE)
}
\arguments{
\item{model_output}{List. Model output generated by \code{run_model()}.}

\item{retain_csv}{Logical Should the Stan csv files be retained. Defaults to
TRUE if user calls function directly. However, when this function is called
internally by \code{run_model} this is set to FALSE.}

\item{path}{Character. Optional file path to use for saved data. Defaults to
the file path used for the original run.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}
}
\value{
Nothing. Creates an \code{.rds} file at \code{path}.
}
\description{
This function closely imitates \code{cmdstanr::save_object()} but saves the
entire model output object from \code{run_model()} which contains more details
regarding data preparation (stratification etc.).
}
\details{
Files are saved to \code{path}, or if not provided, to the original location of
the Stan model run files (if the original files exist).
}
\examples{
# By default, the model is saved as an RDS file during `run_model()`

# But you can also deliberately save the file (here with an example model)
save_model_run(pacific_wren_model, path = "my_model.rds")

# Clean up
unlink("my_model.rds")
}
\seealso{
Other modelling functions: 
\code{\link{copy_model_file}()},
\code{\link{run_model}()}
}
\concept{modelling functions}
