% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_notes}
\alias{species_notes}
\title{Species notes}
\format{
\subsection{\code{species_notes}}{

A data frame with 6 rows and 5 columns
\itemize{
\item \code{english} - The English name of the species
\item \code{french} - Le nom francais, de l'espece
\item \code{aou} - The unique numerical identification of the species in the BBS database
\item \code{minimum_year} - The suggested first year by which the species observations are consistent
\item \code{warning} - The text warning and explanation delivered by the prepare_data function
}
}
}
\usage{
species_notes
}
\description{
Introduced in version 1.1.2.
List of species and comments returned as warning messages if the time-span of
an analysis is not appropriate: A collection of species and species-complexes
where observations are not consistently recorded through time. List intended to
help flag particular species where some trend analyses may be inappropriate.
}
\examples{
species_notes
}
\keyword{datasets}
