% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify and filter Breeding Bird Survey data}
\usage{
stratify(
  by,
  species,
  strata_custom = NULL,
  combine_species_forms = TRUE,
  release = 2025,
  sample_data = FALSE,
  return_omitted = FALSE,
  quiet = FALSE,
  use_map = TRUE,
  distance_to_strata = NULL
)
}
\arguments{
\item{by}{Character. Stratification type. Either an established type, one of
"bbs", prov_state", "bcr", "latlong", "bbs_cws", "bbs_usgs", or a custom name
(see \code{strata_custom} for details).}

\item{species}{Character. Bird species of interest. Can be specified by
English, French, or Scientific names, or AOU code. Use \code{search_species()}
for loose matching to find the exact name/code needed.}

\item{strata_custom}{(\code{sf}) Data Frame. Data frame
of modified existing stratification, or a \code{sf} spatial data frame with
polygons defining the custom stratifications. See Details.}

\item{combine_species_forms}{Logical. Whether to combine ambiguous species
forms. Default \code{TRUE}. See Details.}

\item{release}{Numeric. Which yearly release to use, 2022 (including data
through 2021 field season) or 2020 (including data through 2019). Default
2022.}

\item{sample_data}{Logical. Use sample data (just Pacific Wrens). Default
\code{FALSE}.}

\item{return_omitted}{Logical. Whether or not to return a data frame of
route-years which were omitted during stratification as they did not
overlap with any stratum. For checking and troubleshooting. Default
\code{FALSE}.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}

\item{use_map}{Logical. Whether to stratify the BBS counts and routes based
on the spatial location of route starting points relative to the polygons
in the strata map. If one of the \code{bcr_old}, \code{prov_state}, \code{bbs_cws}, or
\code{bbs_usgs} is supplied to the \code{by} argument, the FALSE option allows
for the routes to be stratified based on the route-specific regional
allocation in the BBS database. Default is TRUE.
Ignored if the user supplies a custom stratification, or the \code{bcr}
or \code{bbs} stratifications are supplied to the \code{by} argument.}

\item{distance_to_strata}{numerical. Maximum distance (in meters), within
which routes with starting locations that do not intersect any strata will
be joined to the nearest stratum. This is NULL by default, indicating that
BBS routes with starting locations that do not overlap a stratum will be
omitted. If supplied, any route with a starting location that falls outside
of the supplied strata polygons, but less than \code{distance_to_strata} meters
from the boundary of at least one of the strata polygons, will be treated
as if they overlap the nearest stratum. This argument is particularly
useful to include routes with starting locations on or near the coast or
shoreline of large lakes, which may be otherwise excluded due to plotting
errors. For example, the map associated with the standard \code{bbs}
stratification excludes 3,877 surveys on 72 routes when this argument is
NULL. All of these 72 routes have starting locations on the coasts.
Setting this argument to 3000 (any route within 3 km of a polygon) ensures
all of these coastal routes are included.}
}
\value{
List of (meta) data.
\itemize{
\item \code{meta_data} - meta data defining the analysis
\item \code{meta_strata} - data frame listing strata names and area for
all strata relevant to the data (i.e. some may have been removed due to
lack of count data). Contains at least \code{strata_name} (the label of the
stratum), and \code{area_sq_km} (area of the stratum).
\item \code{birds_strata} - data frame of stratified count-level data filtered by
species
\item \code{routes_strata} - data frame of stratified route-level data filtered by
species
}
}
\description{
Assign count data to strata and filter by species of interest. Routes are
assigned to strata based on their geographic location and the stratification
specified by the user. Species are filtered by matching English, French or
Scientific names to those in the BBS species data (see \code{search_species()} for
a flexible search to identify correct species names).
}
\details{
To define a custom subset of an existing stratification, specify the
stratification in \code{by} (e.g., "bbs_cws") and then supply a subset of
\code{bbs_strata[["bbc_cws"]]} to \code{strata_custom} (see examples).

To define a completely new custom stratification, specify the name you
would like use in \code{by} (e.g., "east_west_divide") and then supply a spatial
data frame with polygons identifying the different strata to
\code{strata_custom}. Note that this data must have a column called exactly
\code{strata_name} which names all the strata contained (see examples). This
column must also be of class character not numeric or factor.

If \code{combine_species_forms} is \code{TRUE} (default), species with multiple forms
(e.g., "unid. Dusky Grouse / Sooty Grouse") are included in overall species
groupings (i.e., "unid." are combined with "Dusky Grouse" and "Sooty
Grouse" into "Blue Grouse (Dusky/Sooty)"). If the user wishes to keep the
forms separate, \code{combine_species_forms} can be set to \code{FALSE}. See the data
frame \code{species_forms}, for which species are set to be combined with which
other species.

See \code{vignette("stratification", package = "bbsBayes2")} and the article
\href{https://bbsBayes.github.io/bbsBayes2/articles/custom_stratification.html}{custom stratification}
for more details.
}
\examples{
# Sample Data - USGS BBS strata
s <- stratify(by = "bbs_usgs", sample_data = TRUE)

\dontshow{if (have_bbs_data()) withAutoprint(\{ # examplesIf}
# Full data - species and stratification
# Use `search_species()` to get correct species name

# Stratify by CWS BBS strata
s <- stratify(by = "bbs_cws", species = "Common Loon",
              use_map = FALSE)

# Use use English, French, Scientific, or AOU codes for species names
s <- stratify(by = "bbs_cws", species = "Plongeon huard",
              use_map = FALSE)
s <- stratify(by = "bbs_cws", species = 70,
              use_map = FALSE)
s <- stratify(by = "bbs_cws", species = "Gavia immer",
              use_map = FALSE)

# Stratify by updated (2025) Bird Conservation Regions
s <- stratify(by = "bcr", species = "Great Horned Owl",
              distance_to_strata = 4000)

# Stratify by former CWS BBS strata
s <- stratify(by = "bbs_cws", species = "Canada Jay",
              use_map = FALSE)

# Stratify by State/Province/Territory only
s <- stratify(by = "prov_state", species = "Common Loon",
              use_map = FALSE)


# Stratify by blocks of 1 degree of latitude X 1 degree of longitude
s <- stratify(by = "latlong", species = "Snowy Owl",
              use_map = FALSE)

# Check routes omitted by stratification
s <- stratify(by = "latlong", species = "Snowy Owl", return_omitted = TRUE)
s[["routes_omitted"]]

# Use combined or non-combined species forms

search_species("Sooty grouse")
s <- stratify(by = "bbs_usgs", species = "Blue Grouse (Dusky/Sooty)",
              use_map = FALSE)
nrow(s$birds_strata) # Contains all Dusky, Sooty and unidentified

search_species("Sooty grouse", combine_species_forms = FALSE)
s <- stratify(by = "bbs_usgs", species = "unid. Dusky Grouse / Sooty Grouse",
              combine_species_forms = FALSE,
              use_map = FALSE)
nrow(s$birds_strata) # Contains *only* unidentified


# Stratify by a subset of an existing stratification
library(dplyr)
my_cws <- filter(bbs_strata[["bbs_cws"]], country_code == "CA")
s <- stratify(by = "bbs_cws", strata_custom = my_cws, species = "Snowy Owl",
              use_map = FALSE)

my_bcr <- filter(bbs_strata[["bcr_old"]], strata_name == "BCR8")
s <- stratify(by = "bcr_old", strata_custom = my_bcr,
              species = "Yellow-rumped Warbler (all forms)",
              use_map = FALSE)

# Stratify by Custom stratification, using sf map object
# e.g. with WBPHS stratum boundaries 2019
# available: https://ecos.fws.gov/ServCat/Reference/Profile/142628

\dontrun{
map <- sf::read_sf("../WBPHS_Stratum_Boundaries_2019") \%>\%
  rename(strata_name = STRAT) # stratify expects this column

s <- stratify(by = "WBPHS_2019", strata_map = map)
}


\dontshow{\}) # examplesIf}
}
\seealso{
Other Data prep functions: 
\code{\link{prepare_data}()},
\code{\link{prepare_model}()},
\code{\link{prepare_spatial}()}
}
\concept{Data prep functions}
