% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_map}
\alias{plot_map}
\title{Generate a map of trends by strata}
\usage{
plot_map(
  trends,
  slope = FALSE,
  title = TRUE,
  alternate_column = NULL,
  col_viridis = FALSE,
  col_ebird = FALSE,
  strata_custom = NULL,
  zoom_range = TRUE
)
}
\arguments{
\item{trends}{List. Trends generated by \code{generate_trends()}.}

\item{slope}{Logical. Whether or not to map values of the alternative trend
metric (slope of a log-linear regression) if \code{slope = TRUE} was used in
\code{generate_trends()},  through the annual indices. Default \code{FALSE}.}

\item{title}{Logical. Whether or not to include a title with species. Default
\code{TRUE}.}

\item{alternate_column}{Character, Optional name of numerical column in
trends dataframe to plot. If one of the columns with "trend" in the title,
(e.g., trend_q_0.05 then the colour scheme and breaks will match those
used in the default trend maps)}

\item{col_viridis}{Logical. Should the colour-blind-friendly "viridis"
palette be used. Default \code{FALSE}.}

\item{col_ebird}{Logical, Alternative colour palette for trend values, based
on the colour palette used for eBird status and trend maps.}

\item{strata_custom}{(\code{sf}) Data Frame. Data frame
of modified existing stratification, or a \code{sf} spatial data frame with
polygons defining the custom stratifications. See details on strata_custom
in \code{stratify()}.}

\item{zoom_range}{Logical. When TRUE (default) zooms into region of the map
where trend data are available. If FALSE, map extends out to cover all of
the stratification map. Zoom-in uses \code{ggplot2::coord_sf()}}
}
\value{
a ggplot2 plot
}
\description{
\code{plot_map()} allows you to generate a colour-coded map of the percent
change in species trends for each strata.
}
\examples{
# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices
i <- generate_indices(pacific_wren_model)

# Now generate trends
t <- generate_trends(i, slope = TRUE)

# Generate the map (without slope trends)
plot_map(t)

# Generate the map (with slope trends)
plot_map(t, slope = TRUE)

# Viridis
plot_map(t, col_viridis = TRUE)

# Generate a map (with alternate column - lower 95\% Credible limit)
plot_map(t, alternate_column = "trend_q_0.05")

}
\seealso{
Other indices and trends functions: 
\code{\link{generate_indices}()},
\code{\link{generate_trends}()},
\code{\link{plot_geofacet}()},
\code{\link{plot_indices}()}
}
\concept{indices and trends functions}
